refactor: extract completion logic to module
